Create map sprite entities

+++ src/Constants.h
@@ -5,5 +5,5 @@ static const int SCREEN_HEIGHT = 180;
 static const int RENDER_SCALE = 2;
 static const int TILE_SIZE = 16;
 static const float PLAYER_SPEED = 80.0f; // pixels per second
-static const int PLAYER_WIDTH = 32;
-static const int PLAYER_HEIGHT = 32;
+static const float PLAYER_WIDTH = 32;
+static const float PLAYER_HEIGHT = 32;
